CaD 1.1 dev 1 part 2
In this dev build, multiple bugs have been fixed and the code has been simplified. A new feature allows multiple cinematics to run simultaneously using a new cinematic resource called DataCinematic. Now, most changes made to inspector variables in the editor are reflected in real time on the nodes (not all yet, but the majority; those that don’t update are due to issues with refreshing _get_property_list()).

From now on, most upcoming commits will be in English (I’ll try to make it that way).